+++ dad-is-great-in-C/m_sort.cpp
@@ -4,22 +4,22 @@
 #include <iostream>
 #include <random>
 #include <ratio>
-#include <stddef.h>
 #include <vector>
 
-static void merge(std::vector<unsigned int> arr, size_t left, size_t mid,
-                  size_t right) {
-  size_t rt = mid - left + 1, lt = right - mid;
-  int *left_arr = new int[rt],
-      *right_arr = new int[lt]; // создаём два новых массива
+static void merge(std::vector<unsigned int> arr, unsigned long long left,
+                  unsigned long long mid, unsigned long long right) {
+  unsigned long long rt = mid - left + 1, lt = right - mid;
+  unsigned int *left_arr = new unsigned int[rt],
+               *right_arr = new unsigned int[lt]; // создаём два новых массива
 
-  for (size_t i = 0; i < rt; i++)
+  for (unsigned long long i = 0; i < rt; i++)
     left_arr[i] = arr[left + i]; // заполняем их
-  for (size_t i = 0; i < lt; i++)
+  for (unsigned long long i = 0; i < lt; i++)
     right_arr[i] = arr[mid + 1 + i];
 
-  size_t i_right = 0, i_left = 0,
-         i_merged = left; // индексы, по которым мы идём вдоль массивов
+  unsigned long long i_right = 0, i_left = 0,
+                     i_merged =
+                         left; // индексы, по которым мы идём вдоль массивов
 
   while (i_right < rt and i_left < lt) { // слияние
     if (left_arr[i_right] <= right_arr[i_left]) {
@@ -48,19 +48,19 @@ static void merge(std::vector<unsigned int> arr, size_t left, size_t mid,
   delete[] right_arr;
 }
 
-static void merge_sort(std::vector<unsigned int> arr, size_t begin,
-                       size_t end) {
+static void merge_sort(std::vector<unsigned int> arr, unsigned long long begin,
+                       unsigned long long end) {
   if (begin >= end)
     return; // раздел до единичных
 
-  size_t mid = (end + begin) / 2;
+  unsigned long long mid = (end + begin) / 2;
   merge_sort(arr, begin, mid);
   merge_sort(arr, mid + 1, end);
   merge(arr, begin, mid, end);
 }
 
 int main() {
-  size_t n;
+  unsigned long long n;
   std::cin >> n;
   auto st = std::chrono::system_clock::now();
 
@@ -69,7 +69,7 @@ int main() {
   std::mt19937 gen(rd());
   std::uniform_int_distribution<std::mt19937::result_type> dist(1, 1000000);
 
-  for (size_t i = 0; i < n; i++)
+  for (unsigned long long i = 0; i < n; i++)
     a[i] = dist(gen);
   merge_sort(a, 0, n - 1);
