+++ dad-is-great-in-C/s_sort.cpp
@@ -4,11 +4,10 @@
 #include <iostream>
 #include <random>
 #include <ratio>
-#include <stddef.h>
 #include <vector>
 
 int main() {
-  size_t n;
+  unsigned long long n;
   std::cin >> n;
   auto st = std::chrono::system_clock::now();
 
@@ -16,12 +15,12 @@ int main() {
   std::random_device rd;
   std::mt19937 gen(rd());
   std::uniform_int_distribution<std::mt19937::result_type> dist(1, 1000000);
-  for (size_t i = 0; i < n; i++)
+  for (unsigned long long i = 0; i < n; i++)
     a[i] = dist(gen);
 
-  for (size_t i = 0; i < n - 1; i++) {
-    size_t i_min = i;
-    for (size_t j = i + 1; j < n; j++) {
+  for (unsigned long long i = 0; i < n - 1; i++) {
+    unsigned long long i_min = i;
+    for (unsigned long long j = i + 1; j < n; j++) {
       if (a[j] < a[i_min])
         i_min = j;
     }
